Major changes since 0.8. Fido now produces an xml format, rather than the Python format used in previous versions.  The information is held in conf/formats.xml.  In addition, there is a conf/format-extensions.xml that holds format definitions that are not (yet) in pronom.  For example, the current release has signatures for the various office open xml sub-formats.
Fido is now using the python cElementTree library to parse, transform, and produce the formats.xml.  This is not as convenient as the lxml libraries, but it is a built-in library since Python 2.5, and does not introduce any installation dependencies.

As a result of these changes, the code has become much more compact and easier to follow.

Signed-off-by: AFarquha <AFarquha@XPLT200596.ad.bl.uk>