Major changes since 0.8. Fido now produces an xml format, rather than the Python format used in previous versions.  Format definintions have moved to conf/formats.xml and conf/format-extensions.xml.  In addition, it is now possible to load additional format definitions using the -xmlformats argument.
Fido is now using the python cElementTree library to parse, transform, and produce the formats.xml.  This is not as convenient as the lxml libraries, but it is a built-in library since Python 2.5, and does not introduce any installation dependencies.

As a result of these changes, the code has become much more compact and easier to follow.  The invocation overhead is (currently) slight higher than in 0.8, but runtime is comparable.

Note that for some situations, 0.9.1  may provide different results from 0.8, which had some corrected signatures applied by the now removed format-fixup.py.  These should return in 0.9.2!