Don't recurse into every container type
1b5698a updated self.container_type() to recognize OLE as an additional
format, but this broke the `-zip` switch. The method was being used to
identify two different categories of formats:

1. Container formats which need to be matched against the PRONOM
   container signatures in order to get more precise matches; and
2. Container formats which can be recursed into via the `-zip` switch in
   order to identify the formats of their contents.

FIDO supports OLE for the former but not the latter, and since OLE is
usually not interesting for its contents, it doesn't make sense to
support recursing into it.

This commit adds a new method which differentiates whether FIDO is
interested in recursing into a format, not merely whether it *is* a
container format, and updates the `-zip` path to check using it.

Fixes #72.