Make fido importable as a module
*Not tested, yet* Take it as an issue with source code attachment ready to merge if error-free.

Moved argument processing and sys.argv checks out of main() into `if __name__ == '__main__'` clause.

Applied dest='other_name' for certain names that have reserved meaning in python or which are to short to be meaningful. `main()` accepts directly passed keyword arguments (in theory they may be positional, but rather don't), so the caller does not need argparse.